GiftCardFixes: Extract gift card example code into ExampleFullAsyncDropInService
