+++ example-app/src/main/java/com/adyen/checkout/example/service/ExampleAsyncDropInService.kt
@@ -11,19 +11,12 @@ package com.adyen.checkout.example.service
 import com.adyen.checkout.card.CardComponentState
 import com.adyen.checkout.components.ActionComponentData
 import com.adyen.checkout.components.PaymentComponentState
-import com.adyen.checkout.components.model.payments.request.OrderRequest
-import com.adyen.checkout.components.model.payments.request.PaymentMethodDetails
 import com.adyen.checkout.components.model.payments.response.Action
-import com.adyen.checkout.components.model.payments.response.BalanceResult
-import com.adyen.checkout.components.model.payments.response.OrderResponse
 import com.adyen.checkout.core.log.LogUtil
 import com.adyen.checkout.core.log.Logger
-import com.adyen.checkout.core.model.getStringOrNull
 import com.adyen.checkout.core.model.toStringPretty
-import com.adyen.checkout.dropin.service.BalanceDropInServiceResult
 import com.adyen.checkout.dropin.service.DropInService
 import com.adyen.checkout.dropin.service.DropInServiceResult
-import com.adyen.checkout.dropin.service.OrderDropInServiceResult
 import com.adyen.checkout.example.data.api.model.paymentsRequest.AdditionalData
 import com.adyen.checkout.example.data.storage.KeyValueStorage
 import com.adyen.checkout.example.repositories.paymentMethods.PaymentsRepository
@@ -41,8 +34,10 @@ import org.json.JSONObject
 /**
  * This is just an example on how to make networkModule calls on the [DropInService].
  * You should make the calls to your own servers and have additional data or processing if necessary.
+ *
+ * This class implements [onPaymentsCallRequested] and [onDetailsCallRequested] which provide more
+ * freedom in handling the API calls, managing threads and checking component states.
  */
-@Suppress("TooManyFunctions")
 @AndroidEntryPoint
 class ExampleAsyncDropInService : DropInService() {
 
@@ -82,7 +77,7 @@ class ExampleAsyncDropInService : DropInService() {
             val requestBody = paymentRequest.toString().toRequestBody(CONTENT_TYPE)
             val response = paymentsRepository.paymentsRequestAsync(requestBody)
 
-            val result = handleResponse(response) ?: return@launch
+            val result = handleResponse(response)
             sendResult(result)
         }
     }
@@ -105,12 +100,12 @@ class ExampleAsyncDropInService : DropInService() {
             val requestBody = actionComponentJson.toString().toRequestBody(CONTENT_TYPE)
             val response = paymentsRepository.detailsRequestAsync(requestBody)
 
-            val result = handleResponse(response) ?: return@launch
+            val result = handleResponse(response)
             sendResult(result)
         }
     }
 
-    private fun handleResponse(response: ResponseBody?): DropInServiceResult? {
+    private fun handleResponse(response: ResponseBody?): DropInServiceResult {
         val jsonResponse = if (response == null) null else JSONObject(response.string())
         return when {
             jsonResponse == null -> {
@@ -121,12 +116,6 @@ class ExampleAsyncDropInService : DropInService() {
                 Logger.d(TAG, "Received action")
                 val action = Action.SERIALIZER.deserialize(jsonResponse.getJSONObject("action"))
                 DropInServiceResult.Action(action)
-            }
-            isNonFullyPaidOrder(jsonResponse) -> {
-                Logger.d(TAG, "Received a non fully paid order")
-                val order = getOrderFromResponse(jsonResponse)
-                fetchPaymentMethods(order)
-                null
             }
             else -> {
                 Logger.d(TAG, "Final result - ${jsonResponse.toStringPretty()}")
@@ -142,137 +131,5 @@ class ExampleAsyncDropInService : DropInService() {
 
     private fun isAction(jsonResponse: JSONObject): Boolean {
         return jsonResponse.has("action")
-    }
-
-    private fun isNonFullyPaidOrder(jsonResponse: JSONObject): Boolean {
-        return jsonResponse.has("order") && getOrderFromResponse(jsonResponse).remainingAmount?.value ?: 0 > 0
-    }
-
-    private fun getOrderFromResponse(jsonResponse: JSONObject): OrderResponse {
-        val orderJSON = jsonResponse.getJSONObject("order")
-        return OrderResponse.SERIALIZER.deserialize(orderJSON)
-    }
-
-    private fun fetchPaymentMethods(order: OrderResponse? = null) {
-        Logger.d(TAG, "fetchPaymentMethods")
-        launch(Dispatchers.IO) {
-            val orderRequest = if (order == null) null else OrderRequest(
-                pspReference = order.pspReference,
-                orderData = order.orderData
-            )
-            val paymentMethodRequest = getPaymentMethodRequest(keyValueStorage, orderRequest)
-            val paymentMethods = paymentsRepository.getPaymentMethods(paymentMethodRequest)
-            val result = if (paymentMethods != null) {
-                DropInServiceResult.Update(paymentMethods, order)
-            } else {
-                Logger.e(TAG, "FAILED")
-                DropInServiceResult.Error(reason = "IOException")
-            }
-            sendResult(result)
-        }
-    }
-
-    override fun checkBalance(paymentMethodData: PaymentMethodDetails) {
-        launch(Dispatchers.IO) {
-            Logger.d(TAG, "checkBalance")
-
-            val paymentMethodJson = PaymentMethodDetails.SERIALIZER.serialize(paymentMethodData)
-            Logger.v(TAG, "paymentMethods/balance/ - ${paymentMethodJson.toStringPretty()}")
-
-            val paymentRequest = createBalanceRequest(
-                paymentMethodJson,
-                keyValueStorage.getMerchantAccount()
-            )
-
-            val requestBody = paymentRequest.toString().toRequestBody(CONTENT_TYPE)
-            val response = paymentsRepository.balanceRequestAsync(requestBody)
-            val result = handleBalanceResponse(response)
-            sendBalanceResult(result)
-        }
-    }
-
-    private fun handleBalanceResponse(response: ResponseBody?): BalanceDropInServiceResult {
-        return if (response != null) {
-            val balanceJson = response.string()
-            val jsonResponse = JSONObject(balanceJson)
-            val resultCode = jsonResponse.getStringOrNull("resultCode")
-            when (resultCode) {
-                "Success" -> BalanceDropInServiceResult.Balance(BalanceResult.SERIALIZER.deserialize(jsonResponse))
-                "NotEnoughBalance" -> BalanceDropInServiceResult.Error(reason = "Not enough balance", dismissDropIn = false)
-                else -> BalanceDropInServiceResult.Error(reason = resultCode, dismissDropIn = false)
-            }
-        } else {
-            Logger.e(TAG, "FAILED")
-            BalanceDropInServiceResult.Error(reason = "IOException")
-        }
-    }
-
-    override fun createOrder() {
-        launch(Dispatchers.IO) {
-            Logger.d(TAG, "createOrder")
-
-            val paymentRequest = createOrderRequest(
-                keyValueStorage.getAmount(),
-                keyValueStorage.getMerchantAccount()
-            )
-
-            val requestBody = paymentRequest.toString().toRequestBody(CONTENT_TYPE)
-            val response = paymentsRepository.createOrderAsync(requestBody)
-
-            val result = handleOrderResponse(response)
-            sendOrderResult(result)
-        }
-    }
-
-    @Suppress("NestedBlockDepth")
-    private fun handleOrderResponse(response: ResponseBody?): OrderDropInServiceResult {
-        return if (response != null) {
-            val orderJson = response.string()
-            val jsonResponse = JSONObject(orderJson)
-            val resultCode = jsonResponse.getStringOrNull("resultCode")
-            when (resultCode) {
-                "Success" -> OrderDropInServiceResult.OrderCreated(OrderResponse.SERIALIZER.deserialize(jsonResponse))
-                else -> OrderDropInServiceResult.Error(reason = resultCode, dismissDropIn = false)
-            }
-        } else {
-            Logger.e(TAG, "FAILED")
-            OrderDropInServiceResult.Error(reason = "IOException")
-        }
-    }
-
-    override fun cancelOrder(order: OrderRequest, shouldUpdatePaymentMethods: Boolean) {
-        launch(Dispatchers.IO) {
-            Logger.d(TAG, "cancelOrder")
-            val orderJson = OrderRequest.SERIALIZER.serialize(order)
-            val cancelOrderRequest = createCancelOrderRequest(
-                orderJson,
-                keyValueStorage.getMerchantAccount()
-            )
-            val requestBody = cancelOrderRequest.toString().toRequestBody(CONTENT_TYPE)
-            val response = paymentsRepository.cancelOrderAsync(requestBody)
-
-            val result = handleCancelOrderResponse(response, shouldUpdatePaymentMethods) ?: return@launch
-            sendResult(result)
-        }
-    }
-
-    @Suppress("NestedBlockDepth")
-    private fun handleCancelOrderResponse(response: ResponseBody?, shouldUpdatePaymentMethods: Boolean): DropInServiceResult? {
-        return if (response != null) {
-            val orderJson = response.string()
-            val jsonResponse = JSONObject(orderJson)
-            Logger.v(TAG, "cancelOrder response - ${jsonResponse.toStringPretty()}")
-            val resultCode = jsonResponse.getStringOrNull("resultCode")
-            when (resultCode) {
-                "Received" -> {
-                    if (shouldUpdatePaymentMethods) fetchPaymentMethods()
-                    null
-                }
-                else -> DropInServiceResult.Error(reason = resultCode, dismissDropIn = false)
-            }
-        } else {
-            Logger.e(TAG, "FAILED")
-            DropInServiceResult.Error(reason = "IOException")
-        }
     }
 }
